Shared cloud sync test assets
(cherry picked from commit 106b4aad6c61b4697debe34d5c89eed530a79e69)
